upload-1187 integrate with upload server
