--- tus/file-hooks/post-finish-go/go.mod
@@ -1,15 +1,0 @@
-module post-finish-go
-
-go 1.20
-
-require (
-	github.com/Azure/azure-sdk-for-go/sdk/appconfig/azappconfig v0.3.2
-	github.com/joho/godotenv v1.5.1
-)
-
-require (
-	github.com/Azure/azure-sdk-for-go/sdk/azcore v1.0.0 // indirect
-	github.com/Azure/azure-sdk-for-go/sdk/internal v1.0.0 // indirect
-	golang.org/x/net v0.0.0-20220425223048-2871e0cb64e4 // indirect
-	golang.org/x/text v0.3.7 // indirect
-)
